refactor: extract workspace core utilities
